Predict on the latest model (#39)
This patch adds a feature of calling "latest" pushed model.

+++ knuckle/storage/local.py
@@ -2,19 +2,22 @@ import asyncio
 import concurrent.futures
 import io
 import logging
+import operator
 import pathlib
 import shutil
 import tinydb
 import typing
-import uuid
 
 import knuckle.errors
 import knuckle.logging
-import knuckle.model
-import knuckle.storage.meta
 
 from knuckle import asynclib
-from knuckle.storage.meta import query_by_name_and_tag, query_by_id
+from knuckle import model
+from knuckle import signal
+from knuckle.storage import metadata
+from knuckle.storage.metadata import (query_by_name,
+                                      query_by_name_and_tag,
+                                      query_by_id)
 
 
 class FileSystem:
@@ -27,8 +30,8 @@ class FileSystem:
     @classmethod
     def new(cls,
             path: pathlib.Path,
-            meta: knuckle.storage.meta.DB,
-            loader: knuckle.model.Loader,
+            meta: metadata.DB,
+            loader: model.Loader,
             logger: logging.Logger=knuckle.logging.internal_logger):
 
         self = cls()
@@ -39,21 +42,32 @@ class FileSystem:
         self.loader = loader
         self.models_path = path.joinpath("models")
 
+        self._on_delete = signal.Signal()
+        self._on_save = signal.Signal()
+
         self.models_path.mkdir(parents=True, exist_ok=True)
         self.executor = concurrent.futures.ThreadPoolExecutor()
 
         return self
 
-    def _new_model(self, document: typing.Dict) -> knuckle.model.Model:
+    @property
+    def on_delete(self):
+        return self._on_delete
+
+    @property
+    def on_save(self):
+        return self._on_save
+
+    def build_model_from_document(self, document: typing.Dict) -> model.Model:
         path = self.models_path.joinpath(document["id"])
-        return knuckle.model.Model(path=path, loader=self.loader, **document)
+        return model.Model(path=path, loader=self.loader, **document)
 
     def await_in_thread(self, task: typing.Coroutine):
         """Run the given function within an instance executor."""
         loop = asyncio.get_event_loop()
         return loop.run_in_executor(self.executor, asynclib.run, task)
 
-    async def all(self) -> typing.Sequence[knuckle.model.Model]:
+    async def all(self) -> typing.Sequence[model.Model]:
         """List available models and their tags.
 
         The method returns a list of not loaded models, therefore before using
@@ -61,40 +75,54 @@ class FileSystem:
         """
         for document in await self.meta.all():
             path = self.models_path.joinpath(document["id"])
-            m = knuckle.model.Model(path=path, loader=self.loader, **document)
+            m = model.Model(path=path, loader=self.loader, **document)
             yield m
 
+    async def save_to_meta(self, m: model.Model) -> None:
+        async with self.meta.write_locked() as meta:
+            if await meta.get(query_by_name_and_tag(m.name, m.tag)):
+                self.logger.debug("Model %s already exists", m)
+
+                raise knuckle.errors.DuplicateError(m.name, m.tag)
+
+            # Insert the model metadata, and update the latest model link.
+            await meta.insert(m.to_dict())
+            await self.on_save.send(m)
+
+            # Since the saving is happening right now, the latest model
+            # will obviously be the current one.
+            latest = m.copy()
+
+            latest.tag = model.Tag.Latest.value
+            latest.id = m.id
+
+            latest_query = query_by_name_and_tag(latest.name, latest.tag)
+            await meta.upsert(latest.to_dict(), latest_query)
+            await self.on_save.send(latest)
+
     async def save(self, name: str, tag: str,
-                   model: io.IOBase) -> knuckle.model.Model:
+                   stream: io.IOBase) -> model.Model:
         """Save the model into the local storage.
 
-        Extracts the TAR archive into the data root directory.
+        Extracts the TAR archive into the data directory.
         """
-        model_id = uuid.uuid4()
-        model_path = self.models_path.joinpath(model_id.hex)
+        m = model.Model.new(name, tag, self.models_path, self.loader)
 
         try:
-            task = asynclib.extract_tar(fileobj=model, dest=model_path)
+            task = asynclib.extract_tar(fileobj=stream, dest=m.path)
             await self.await_in_thread(task)
 
             # Now load the model into the memory, to pass all validations.
             self.logger.debug("Ensuring model has correct format")
 
-            m = knuckle.model.Model(model_id, name, tag,
-                                    model_path, self.loader)
-            m = await self.await_in_thread(asyncio.coroutine(m.load)())
+            task = asyncio.coroutine(m.load)()
+            m = await self.await_in_thread(task)
 
-            async with self.meta.write_locked() as meta:
-                if await meta.get(query_by_name_and_tag(name, tag)):
-                    self.logger.debug("Model %s already exists", m)
-                    raise knuckle.errors.DuplicateError(name, tag)
-
-                # Insert the model metadata only on the last step.
-                await meta.insert(m.to_dict())
+            await self.save_to_meta(m)
 
             # Model successfully loaded, so now it can be moved to the original
             # data root directory.
-            self.logger.info("Pushing model %s to %s", m, model_path)
+            self.logger.info("Pushing model %s to %s", m, m.path)
             return m
 
         except Exception as e:
@@ -102,33 +130,59 @@ class FileSystem:
             # and ensure the metadata database does not store any information.
             #
             # The caller have to ensure atomicity of this operation.
-            await self.meta.remove(query_by_id(model_id))
+            await self.meta.remove(query_by_id(m.id))
 
-            task = asynclib.remove_dir(model_path, ignore_errors=True)
+            task = asynclib.remove_dir(m.path, ignore_errors=True)
             await self.await_in_thread(task)
             raise e
 
+    async def delete_from_meta(self, name: str, tag: str) -> model.Model:
+        # Model found, remove metadata from the database.
+        async with self.meta.write_locked() as meta:
+            m = await self.load_from_meta(name, tag)
+
+            await meta.remove(query_by_id(m.id))
+            await self.on_delete.send(m.name, m.tag)
+
+            # Remove the "latest" model link.
+            query = query_by_name_and_tag(m.name, model.Tag.Latest.value)
+            await meta.remove(query)
+
+            # Retrieve a new "latest" model.
+            key = operator.itemgetter("created_at")
+            document = await meta.latest(query_by_name(m.name), key)
+
+            latest = self.build_model_from_document(document)
+            latest.tag = model.Tag.Latest.value
+
+            await meta.insert(latest.to_dict())
+            await self.on_delete.send(m.name, model.Tag.Latest.value)
+            await self.on_save.send(latest)
+        return m
+
     async def delete(self, name: str, tag: str) -> None:
         """Remove model with the given name and tag."""
-        try:
-            # Model found, remove metadata from the database.
-            m = await self._load(name, tag)
-            await self.meta.remove(query_by_id(m.id))
+        if tag == model.Tag.Latest.value:
+            raise model.NotFoundError(name, tag)
 
-            # Remove the model data from the storage.
+        try:
+            # Remove the model from the metadata database.
+            m = await self.delete_from_meta(name, tag)
+
+            # Remove the model data from the file system.
             await self.await_in_thread(asynclib.remove_dir(m.path))
 
             self.logger.info("Removed model %s:%s", name, tag)
         except FileNotFoundError:
             raise knuckle.errors.NotFoundError(name, tag)
 
-    async def _load(self, name: str, tag: str):
+    async def load_from_meta(self, name: str, tag: str):
         document = await self.meta.get(query_by_name_and_tag(name, tag))
         if not document:
             raise knuckle.errors.NotFoundError(name, tag)
-        return self._new_model(document)
+        return self.build_model_from_document(document)
 
-    async def load(self, name: str, tag: str) -> knuckle.model.Model:
+    async def load(self, name: str, tag: str) -> model.Model:
         """Load model with the given name and tag."""
-        m = await self._load(name, tag)
+        m = await self.load_from_meta(name, tag)
         return await self.await_in_thread(asyncio.coroutine(m.load)())
